Prevent ThemeMgr from recreating window handles during form creation
Window handle recreation during form creation causes an inversion of the component's CreateWnd/Loaded calling order, which causes bugs with some components. Such a bug was seen with the lvIMClients listview component in WMP Plus's settings. If the listview was sorted and its window handle recreation was triggered by opening the ATF editor, the listview item's check and enabled states were restored in the wrong order.
Also added checks to TExtListView and TExtChkListView to make sure it won't go unnoticed anymore when check/enabled states would be restored in the wrong order.