Revert back to declaring the SID_STopLevelBrowser constant as a TGUID
This was changed to a string declaration as part of commit 81cdc057890096afb60093842450ee5537897e30 (to go back to the original ShlObj2.pas as much as possible), but this breaks the SmartClose build.
Note that most of the other SID_ constants must remain being declared as string since these are used in interface declarations, where Delphi only accepts string constants.