Fix possible nil pointer dereference in TCustomForm.SetWindowToMonitor
This occurs when the application is first started minimized, and is only being restored after Windows has detected the addition/removal of a secondary monitor.
It was fixed by porting over a piece of code from the Delphi 7 version of TCustomForm.GetMonitor.
Additionally, I made the code in TCustomForm.SetWindowToMonitor more robust to guard against possible nil pointers in the future.