Fix possible infinite loop of TIconMessage.AdjustBounds calls
This method calls TWinControl.SetBounds, which in turn can cause TIconMessage.SetBounds to be called again, resulting in the CM_ADJUSTBOUNDS message being posted to the message queue over and over again.
This was seen with the msgCopyNotes control of TTagEditorForm when the control has an empty Caption, right after the form was shown.