TCustomTab95Control: disable alignment while the control is being loaded from a form stream
This is a fix for Delphi incorrectly 'thinking' that the tab control was resized during loading (due to the 'Images' property being set quite late in the loading phase). The result was that Delphi would align (i.e. resize) some of the tab's child controls incorrectly. The effect wasn't immediately noticeable during program execution (since it usually resulted in the resize by just 1 pixel), but much more so while using the Delphi IDE, since the affected controls were being resized incrementally on every load/save iteration of the form.
Note: the same bug also occurs with Delphi's own TCustomTabControl in the ComCtrls unit, but for now, only the TCustomTab95Control was fixed.